+++ src/core/crypto/ripemd160.h
@@ -15,6 +15,8 @@ public:
     void update(const u8 * data, size_t length);
     void update(const str &data);
     str digest();
+
+    static str fast(const str& input);
 private:
     u32 key[RMD160_BLOCKWORDS]{};
     u32 iv[RMD160_HASHWORDS]{};